# Add struct field + associated tags caching