--- a/util.go
+++ b/util.go
@@ -247,11 +247,62 @@ func panicIf(err error) {
 	}
 }
 
+func (v *Validate) parseStruct(topStruct reflect.Type, sName string) *cachedStruct {
+
+	s := &cachedStruct{Name: sName, fields: map[int]cachedField{}}
+
+	numFields := topStruct.NumField()
+
+	var fld reflect.StructField
+	var tag string
+	var customName string
+
+	for i := 0; i < numFields; i++ {
+
+		fld = topStruct.Field(i)
+
+		if len(fld.PkgPath) != 0 {
+			continue
+		}
+
+		tag = fld.Tag.Get(v.tagName)
+
+		if tag == skipValidationTag {
+			continue
+		}
+
+		customName = fld.Name
+		if len(v.fieldNameTag) != 0 {
+
+			name := strings.SplitN(fld.Tag.Get(v.fieldNameTag), ",", 2)[0]
+
+			// dash check is for json "-" (aka skipValidationTag) means don't output in json
+			if name != "" && name != skipValidationTag {
+				customName = name
+			}
+		}
+
+		cTag, ok := v.tagCache.Get(tag)
+		if !ok {
+			cTag = v.parseTags(tag, fld.Name)
+		}
+
+		s.fields[i] = cachedField{Idx: i, Name: fld.Name, AltName: customName, CachedTag: cTag}
+	}
+
+	v.structCache.Set(sName, s)
+
+	return s
+}
+
 func (v *Validate) parseTags(tag, fieldName string) *cachedTag {
 
 	cTag := &cachedTag{}
 
 	v.parseTagsRecursive(cTag, tag, fieldName, blank, false)
+
+	v.tagCache.Set(tag, cTag)
+
 	return cTag
 }